+++ test/bigtest/interactors/action-menu-interactor.js
@@ -13,6 +13,7 @@ export class ActionMenuInteractor {
   selectAllButton = new ButtonInteractor('[data-test-select-all-items-menu-button]');
   deselectAllButton = new ButtonInteractor('[data-test-deselect-all-items-menu-button]');
   loadRecordsButton = new ButtonInteractor('[data-test-load-records]');
+  restoreDefaultButton = new ButtonInteractor('[data-test-restore-default-file-extensions-button]');
   editProfile = new ButtonInteractor('[data-test-edit-item-menu-button]');
   duplicateProfile = new ButtonInteractor('[data-test-duplicate-item-menu-button]');
   deleteProfile = new ButtonInteractor('[data-test-delete-item-menu-button]');