+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * Add caret to the incoming record select dropdown in match profile
 * Upgrade Stripes and all the dependencies to version 3.0.0 (UIDATIMP-422)
 * Rename renewal.json to ongoing.json (UIDATIMP-428)
+* Remove extra action buttons in File extension Settings screens after central component update (UIDATIMP-398)
 
 ### Bugs fixed:
 * Profile Associator lists are empty when the user reloads the page with Profile Edit Form open (UIDATIMP-338)